fix: resolve all 30 MSVC warnings across full debug build
- Add static_cast for narrowing conversions (CryptoLayer, NLFilterGenerator, nat_demo)
- Mark unused parameters with [[maybe_unused]] (main, SyncManager, efsw_test,
  nat_demo, test_sync_logic, test_transfer_manager_coverage)
- Suppress C4459 global variable shadowing in nat_demo (demo callback params)
- Suppress C4996 getenv deprecation warning in main (safe usage in #ifdef _WIN32)
- Comment out unreachable guard.commit() in test_database_advanced
- Mark unused test variables with [[maybe_unused]] in test_kcp_session

Full clean rebuild: 0 errors, 0 warnings across all 5 targets.

+++ src/common/CryptoLayer.cpp
@@ -90,7 +90,7 @@ void CryptoLayer::set_key(const std::string& key_string) {
     ok = ok && (EVP_PKEY_CTX_set1_hkdf_salt(ctx, salt, sizeof(salt) - 1) == 1);
     ok = ok && (EVP_PKEY_CTX_set1_hkdf_key(ctx,
             reinterpret_cast<const unsigned char*>(key_string.c_str()),
-            key_string.length()) == 1);
+            static_cast<int>(key_string.length())) == 1);
     ok = ok && (EVP_PKEY_CTX_add1_hkdf_info(ctx, info, sizeof(info) - 1) == 1);
 
     size_t outlen = AES256_KEY_SIZE;
